Added adjust-mflags patch file
Adjust-mflags.sh under OpenJDK8's hotsomething folder is borked, and requires it be 'fixed' (read: replaced) with an updated bash script. Long story short, they use a bash terminology piece in a regex that was phased out in newer bash regex, so it doesn't work properly. The hotfix isn't my own, it was sourced from a post on stackexchange and if it doesn't work, please let me know (it probably doesn't).