Upgraded download-github to be multi-modular
Uses multiple approaches when trying to obtain github directories.
Instead of just using mv, which can fail if a directory pre-exists, download-github.sh now calls recursive-copy.sh in the event the directory already exists, complete with appropriate clean-up of the folder it copies from.

Because Download-Github.sh might be downloaded by itself (IE without recursive-copy.sh) it retains mv for the case scenarios it's being deployed to a fresh install. Flexibility for everyone!